Update schema and add Procfile

+++ db/schema.rb
@@ -16,8 +16,8 @@ ActiveRecord::Schema.define(version: 2022_01_02_084855) do
   enable_extension "plpgsql"
 
   create_table "tweets", force: :cascade do |t|
-    t.integer "user_id", null: false
-    t.integer "twitter_account_id", null: false
+    t.bigint "user_id", null: false
+    t.bigint "twitter_account_id", null: false
     t.text "body"
     t.datetime "publish_at"
     t.string "tweet_id"
@@ -28,7 +28,7 @@ ActiveRecord::Schema.define(version: 2022_01_02_084855) do
   end
 
   create_table "twitter_accounts", force: :cascade do |t|
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.string "name"
     t.string "username"
     t.string "image"
